+++ pyqgiswps/handlers/storehandler.py
@@ -17,6 +17,7 @@ from datetime import datetime
 from pyqgiswps.executors.logstore import logstore
 
 from .basehandler import BaseHandler
+from .apihandler import RealmController
 
 from typing import (
     Optional,
@@ -42,61 +43,82 @@ class StoreShellMixIn():
     """
 
     # XXX DEPRECATED - will be removed at version 1.9
-    def legacy_resource_list(self, path, uuid):
-        # Legacy json schema for
+    def legacy_resource_list(self, workdir: Path, uuid: str):
+        # Only if admin
+        if self.realm_enabled():
+            raise HTTPError(401)
+        # Legacy json schema
         store_url = f"{self.proxy_url()}store/{uuid}/"
-
+    
         def file_records(rootdir, dirs, files, rootfd):
             root = Path(rootdir)
             for file in files:
                 stat  = os.stat(file, dir_fd=rootfd)
-                name  = (root / file).relative_to(path).as_posix()
+                name  = (root / file).relative_to(workdir).as_posix()
                 yield {
                     'name': name,
                     'content_length': stat.st_size,
-                    'store_url'     : store_url + name,
+                    'store_url'     : f"{store_url}{name}",
                     'display_size'  : _format_size(stat.st_size),
                 }
-        data = [f for args in os.fwalk(path) for f in file_records(*args)]
-        content = { "files": data }       
+        data = [f for args in os.fwalk(workdir) for f in file_records(*args)]
+        self.write_json({"files":data})       
 
-        return content
-
-    async def ls( self, uuid ):
+    async def ls( self, uuid: str):
         """ List all files in workdir
         """
-        path = os.path.join(self._workdir, uuid)
-        if not os.path.isdir(path):
+        workdir = Path(self._workdir, uuid)
+        if not workdir.is_dir():
             raise HTTPError(404, reason="The resource does not exists")
-
 
         if self._legacy:
             # XXX DEPRECATED - will be removed at version 1.9
-            content = self.legacy_resource_list(path, uuid) 
+            self.legacy_resource_list(workdir)
+            return
+
+        store_url = f"{self.proxy_url()}jobs/{uuid}/files"
+
+        if self.realm_enabled():
+
+            job_status = self.application.wpsservice.get_status(uuid)
+            if job_status is None:
+                raise HTTPError(404, reason="The resource does not exists") 
+
+            # Restrict to output files
+            def file_records(files):
+                for name in files:
+                    path = workdir.joinpath(name)
+                    if not path.is_file():
+                        continue
+                    content_type = mimetypes.types_map.get(path.suffix)
+                    size = path.stat().st_size
+                    yield {
+                        'href': f"{store_url}/{name}",
+                        'type': content_type or "application/octet-stream",
+                        'name': name,
+                        'content_length': size,
+                        'display_size'  : _format_size(size),
+                    }
+            data = list(file_records(job_status.get('output_files',[])))
         else:
-            store_url = f"{self.proxy_url()}jobs/{uuid}/"
-            # Open api conformant schema
             def file_records(rootdir, dirs, files, rootfd):
                 root = Path(rootdir)
                 for file in files:
-                    stat  = os.stat(file, dir_fd=rootfd)
-                    name  = (root / file).relative_to(path).as_posix()
-                    content_type = mimetypes.types_map.get(
-                        os.path.splitext(name)[1]
-                    ) or "application/octet-stream"
+                    stat = os.stat(file, dir_fd=rootfd)
+                    path = root.joinpath(file).relative_to(workdir)
+                    content_type = mimetypes.types_map.get(path.suffix)
                     yield {
-                        'href': store_url + name,
-                        'type': content_type,
-                        'name': name,
+                        'href': f"{store_url}/{path}",
+                        'type': content_type or "application/octet-stream",
+                        'name': path.as_posix(),
                         'content_length': stat.st_size,
                         'display_size'  : _format_size(stat.st_size),
                     }
-            data = [f for args in os.fwalk(path) for f in file_records(*args)]
-            content = { "links": data }    
+            data = [f for args in os.fwalk(workdir) for f in file_records(*args)]
+            
+        self.write_json({"links": data})    
 
-        self.write_json(content)                  
-
-    async def dnl( self, uuid, filename):
+    async def dnl( self, uuid, filename, content_type: Optional[str]=None):
         """ Return output file from process working dir
         """
         full_path  = os.path.join(self._workdir, uuid, filename)
@@ -115,7 +137,9 @@ class StoreShellMixIn():
             return
 
         # Set headers
-        content_type = mimetypes.types_map.get(os.path.splitext(full_path)[1]) or "application/octet-stream"
+        if content_type is None:
+            content_type = mimetypes.types_map.get(os.path.splitext(full_path)[1]) or "application/octet-stream"
+
         self.set_header("Content-Type", content_type)       
         self.set_header("Etag", etag)
 
@@ -157,9 +181,8 @@ class StoreShellMixIn():
         })
 
 
-class StoreHandler(BaseHandler, StoreShellMixIn):
-    """ Handle store requests
-    """
+class StoreHandlerBase(BaseHandler, StoreShellMixIn, RealmController):
+
     def initialize(self, workdir, chunk_size=65536, legacy: bool=False, ttl=30):
         super().initialize()
         self._chunk_size = chunk_size
@@ -167,6 +190,10 @@ class StoreHandler(BaseHandler, StoreShellMixIn):
         self._legacy = legacy
         self._ttl = ttl
 
+
+class StoreHandler(StoreHandlerBase):
+    """ Handle store requests
+    """
     async def get(self, uuid: str, resource: Optional[str]=None) -> Awaitable[None]:
         """ Handle GET request
         """
@@ -179,12 +206,17 @@ class StoreHandler(BaseHandler, StoreShellMixIn):
         """ Handle put actions
         """
         if not uuid:
-            raise HTTPError(403, reason="Missing job ID")
+            raise HTTPError(400, reason="Missing job ID")
 
         # Get the status for uuid
         job_status = self.application.wpsservice.get_status(uuid)
         if job_status is None:
             raise HTTPError(404, reason="The resource does not exists") 
+
+        if self.realm_enabled():
+            allowed_files = job_status.get('output_files',[])
+            if resource not in allowed_files:
+                raise HTTPError(401)
 
         command = self.get_argument('COMMAND', default="").lower()
         if command == 'geturl':
@@ -193,17 +225,12 @@ class StoreHandler(BaseHandler, StoreShellMixIn):
             raise HTTPError(400, reason=f"Invalid command parameter {command}")
 
 
-class DownloadHandler(BaseHandler, StoreShellMixIn):
+class DownloadHandler(StoreHandlerBase):
     """ The safe store handler work in two time
 
         1. Return a tokenized set of  url with a limited ttl
         2. Handle the url by returning the appropriate ressource
     """
-    def initialize(self, workdir, query=False,  chunk_size=65536):
-        super().initialize()
-        self._chunk_size = chunk_size
-        self._workdir = workdir
-
     async def get(self, token):
         """ Handle GET request
         """
@@ -218,5 +245,20 @@ class DownloadHandler(BaseHandler, StoreShellMixIn):
         if params is None:
             raise HTTPError(403)
         return params['uuid'], params['resource']
+
+
+class LogsHandler(StoreHandlerBase):
+    """ Retrun log text content
+    """
+    async def get(self, job_id) -> Awaitable[None]:
+        """ Handle GET request
+        """
+        # Only admin can see this
+        if self.realm_enabled():
+            raise HTTPError(403)
+
+        await self.dnl(job_id, "processing.log", 
+                       content_type="text/plain; charset=utf-8")
+