+++ tests/unittests/test_complexoutput.py
@@ -10,6 +10,12 @@ from pyqgiswps.inout import (LiteralInput,
 
 from pyqgiswps.ogc.ows import WPS, OWS
 
+
+class OutputContext:
+    def resolve_store_url(self, url: str, as_output: bool=False) -> str:
+        return url
+
+
 def test_complex_output_href():
     """ Test external reference in complex output 
     """
@@ -28,9 +34,9 @@ def test_complex_output_href():
     output.output_format = "application/x-ogc-wms"
     output.url = "http://my.org/external/ref"
 
-    output_elements = output.execute_xml()
+    output_elements = output.execute_xml(OutputContext())
 
-    # Check that <wps:Reference href=...> is not namspaced
+    # Check that <wps:Reference href=...> is not namespaced
     element =  output_elements.xpath('//wps:Reference', namespaces={'wps': "http://www.opengis.net/wps/1.0.0"})
 
     assert len(element) == 1